+++ redisinsight/ui/src/components/side-panels/panels/ai-assistant/components/shared/restart-chat/RestartChat.tsx
@@ -59,4 +59,4 @@ const RestartChat = (props: Props) => {
   )
 }
 
-export default RestartChat
+export default React.memo(RestartChat)
